refactoring moving temp file creation helper into `helper.rb`
